requirements.txt
asyncpg==0.30.0
attrs==25.3.0
black==21.9b0
click==8.2.1
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
py==1.11.0
pytest==6.2.5
python-dotenv==0.16.0
PyYAML==6.0.2
regex==2025.8.29
toml==0.10.2
tomli==1.2.3
typing_extensions==4.15.0
